[WEEK-2] Coin change problem (#8)
Dynamic programming approach to the problem of adding up a change amount using the least amount of coins from given denominations. Commit includes

* Unit testing of the utils

PS: Property testing pending. Brute force solution is not easy to implement in clojure. I also tried a greedy approach based on the criteria to use the coin with largest denomination that had the smallest reminder of the amount. Although this works for a larger set than the direct greedy approach to use the largest denomination, it still failed for some test cases. So not universal, but better than the direct greedy approach.